+++ opentelemetry-lightstep/exe/eventlog-to-lightstep/Main.hs
@@ -6,8 +6,8 @@ import Data.Function
 import qualified Data.Text as T
 import OpenTelemetry.EventlogStreaming_Internal
 import OpenTelemetry.Exporter
-import OpenTelemetry.LightStep.Config
-import OpenTelemetry.LightStep.Exporter
+import OpenTelemetry.Lightstep.Config
+import OpenTelemetry.Lightstep.Exporter
 import System.Clock
 import System.Environment (getArgs, getEnvironment)
 import System.FilePath
@@ -20,18 +20,18 @@ main = do
   args <- getArgs
   case args of
     ["read", path] -> do
-      printf "Sending %s to LightStep...\n" path
+      printf "Sending %s to Lightstep...\n" path
       Just lsConfig <- getEnvConfig
       let service_name = T.pack $ takeBaseName path
-      exporter <- createLightStepSpanExporter lsConfig {lsServiceName = service_name}
+      exporter <- createLightstepSpanExporter lsConfig {lsServiceName = service_name}
       origin_timestamp <- fromIntegral . toNanoSecs <$> getTime Realtime
       withFile path ReadMode (work origin_timestamp exporter)
       shutdown exporter
       putStrLn "\nAll done.\n"
     ("run" : program : "--" : args') -> do
-      printf "Streaming eventlog of %s to LightStep...\n" program
+      printf "Streaming eventlog of %s to Lightstep...\n" program
       Just lsConfig <- getEnvConfig
-      exporter <- createLightStepSpanExporter lsConfig {lsServiceName = T.pack program}
+      exporter <- createLightstepSpanExporter lsConfig {lsServiceName = T.pack program}
       let pipe = program <> "-opentelemetry.pipe"
       runProcess $ proc "mkfifo" [pipe]
       origin_timestamp <- fromIntegral . toNanoSecs <$> getTime Realtime