+++ opentelemetry-lightstep/src/OpenTelemetry/LightStep/Exporter.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 
-module OpenTelemetry.LightStep.Exporter where
+module OpenTelemetry.Lightstep.Exporter where
 
 import Control.Concurrent.Async
 import Control.Concurrent.STM
@@ -14,13 +14,13 @@ import Network.HTTP.Client.TLS
 import Network.HTTP.Types
 import OpenTelemetry.Common
 import OpenTelemetry.Exporter
-import OpenTelemetry.LightStep.Config
+import OpenTelemetry.Lightstep.Config
 import OpenTelemetry.SpanContext
 import OpenTelemetry.ZipkinExporter
 import Text.Printf
 
-createLightStepSpanExporter :: MonadIO m => LightStepConfig -> m (Exporter Span)
-createLightStepSpanExporter cfg@(LightStepConfig {..}) = liftIO do
+createLightstepSpanExporter :: MonadIO m => LightstepConfig -> m (Exporter Span)
+createLightstepSpanExporter cfg@(LightstepConfig {..}) = liftIO do
   let zcfg =
         ZipkinConfig
           { zEndpoint = printf "https://%s:%d/api/v2/spans" lsHostName (fromIntegral lsPort :: Int),