+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ## 0.1.0 -- 2020-03-04
 
-* Add LightStep trace exporter using Zipkin format
+* Add Lightstep trace exporter using Zipkin format
 * Automatically set error=true tag on exceptions
 
 ## 0.0.0.2 -- 2020-02-01